Refactor Income into Handler, add deductions
Incomes were previously tracked as lists of ints. Adding in deductions made it very unclear which incomes were 'total' income and which were 'taxable' income - which also differs at the federal and state levels. A FederalIncomeHandler was then added which computes both 'total' and 'taxable' federal incomes, based on variety of optional income sources. Deductions were also added to that Handler to support itemized deduction and the standard deduction.